+++ weblate/memory/tests.py
@@ -22,8 +22,11 @@ from __future__ import unicode_literals
 from django.test import TestCase
 from django.core.management import call_command
 
+from weblate.lang.models import Language
+from weblate.memory.machine import WeblateMemory
 from weblate.memory.models import Memory
 from weblate.trans.tests.utils import get_test_file
+from weblate.trans.tests.test_checks import MockUnit
 
 
 class MemoryTest(TestCase):
@@ -33,3 +36,24 @@ class MemoryTest(TestCase):
             get_test_file('memory.tmx')
         )
         self.assertEqual(Memory.objects.count(), 2)
+
+    def test_machine(self):
+        Memory.objects.create(
+            source_language=Language.objects.get(code='en'),
+            target_language=Language.objects.get(code='cs'),
+            source='Hello',
+            target='Ahoj',
+            origin='test'
+        )
+        machine_translation = WeblateMemory()
+        self.assertEqual(
+            machine_translation.translate('cs', 'Hello', MockUnit(), None),
+            [
+                {
+                    'quality': 100,
+                    'service': 'Weblate (test)',
+                    'source': 'Hello',
+                    'text': 'Ahoj'
+                },
+            ]
+        )